Abstract LabelScanStoreTest, implemented by Lucene
Separated Lucene specific parts from LuceneLabelScanStoreTest and created an abstract LabelScanStoreTest
that should be used by other LabelScanStore implementations.
