+++ enterprise/ha/src/test/java/org/neo4j/kernel/api/impl/labelscan/LabelScanStoreHaIT.java
@@ -19,18 +19,18 @@
  */
 package org.neo4j.kernel.api.impl.labelscan;
 
-import org.apache.lucene.store.LockObtainFailedException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-import java.util.Arrays;
+import java.util.Collections;
 
 import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Label;
 import org.neo4j.graphdb.Transaction;
 import org.neo4j.graphdb.factory.TestHighlyAvailableGraphDatabaseFactory;
+import org.neo4j.kernel.api.labelscan.LabelScanStore;
 import org.neo4j.kernel.extension.KernelExtensionFactory;
 import org.neo4j.kernel.impl.ha.ClusterManager;
 import org.neo4j.kernel.impl.ha.ClusterManager.ManagedCluster;
@@ -95,7 +95,7 @@ public class LabelScanStoreHaIT
     private enum Labels implements Label
     {
         First,
-        Second;
+        Second
     }
 
     @Rule
@@ -109,7 +109,7 @@ public class LabelScanStoreHaIT
     {
         KernelExtensionFactory<?> testExtension = new LuceneLabelScanStoreExtension( 100, monitor );
         TestHighlyAvailableGraphDatabaseFactory factory = new TestHighlyAvailableGraphDatabaseFactory();
-        factory.addKernelExtensions( Arrays.asList( testExtension ) );
+        factory.addKernelExtensions( Collections.singletonList( testExtension ) );
         ClusterManager clusterManager = new ClusterManager.Builder( testDirectory.directory( "root" ) )
                 .withDbFactory( factory )
                 .withStoreDirInitializer( ( serverId, storeDir ) -> {
@@ -144,7 +144,7 @@ public class LabelScanStoreHaIT
         life.shutdown();
     }
 
-    private static class TestMonitor implements LuceneLabelScanStore.Monitor
+    private static class TestMonitor implements LabelScanStore.Monitor
     {
         private volatile int callsTo_init;
         private volatile int timesRebuiltWithData;
@@ -161,12 +161,12 @@ public class LabelScanStoreHaIT
         }
 
         @Override
-        public void lockedIndex( LockObtainFailedException e )
+        public void lockedIndex( Exception e )
         {
         }
 
         @Override
-        public void corruptedIndex()
+        public void notValidIndex()
         {
         }
